隨機數續修正
git-svn-id: http://www.castiello-tsai.org:8523/svn/L1JTW@124 96aee97a-5c11-ed42-8518-d7f33e96f5f9

+++ src/l1j/server/server/model/DungeonRandom.java
@@ -114,7 +114,7 @@ public class DungeonRandom {
 
 		private NewDungeonRandom(int[] newX, int[] newY, short[] newMapId,
 				int heading) {
-			for (int i = 0; i < 5; i++) {
+			for (byte i = 0; i < 5; i++) {
 				_newX[i] = newX[i];
 				_newY[i] = newY[i];
 				_newMapId[i] = newMapId[i];
